+++ test/unit/last.test.ts
@@ -4,14 +4,17 @@ import {
     dbCount,
     BROADCAST_CHANNEL_BY_TOKEN
 } from '../../';
-import { OPEN_POUCHDB_STORAGE_INSTANCES } from '../../plugins/pouchdb';
+import {
+    OPEN_POUCHDB_STORAGE_INSTANCES,
+    OPEN_POUCH_INSTANCES
+} from '../../plugins/pouchdb';
 import config from './config';
 
 describe('last.test.ts (' + config.storage.name + ')', () => {
     it('ensure every db is cleaned up', () => {
         assert.strictEqual(dbCount(), 0);
     });
-    it('ensure every storage instance is cleaned up', async () => {
+    it('ensure every PouchDB storage instance is cleaned up', async () => {
         try {
             // for performance, we do not await db closing, so it might take some time
             // until everything is closed.
@@ -21,6 +24,18 @@ describe('last.test.ts (' + config.storage.name + ')', () => {
         } catch (err) {
             console.dir(OPEN_POUCHDB_STORAGE_INSTANCES);
             throw new Error('no all storage instances have been closed');
+        }
+    });
+    it('ensure every PouchDB database is removed', async () => {
+        try {
+            // for performance, we do not await db closing, so it might take some time
+            // until everything is closed.
+            await waitUntil(() => {
+                return OPEN_POUCH_INSTANCES.size === 0;
+            }, 5 * 1000, 500);
+        } catch (err) {
+            console.dir(OPEN_POUCH_INSTANCES);
+            throw new Error('no all pouch instances have been closed');
         }
     });
     it('ensure all BroadcastChannels are closed', () => {